+++ package.json
@@ -17,6 +17,7 @@
     "highlight.js": "^11.10.0",
     "marked": "^14.0.0",
     "marked-highlight": "^2.1.4",
+    "pinia": "^2.2.2",
     "vue": "^3.4.35",
     "vue-router": "^4.4.3",
     "vue-toast-notification": "^3.1.2"
